feature/jin-538 Implement material access management: add access types, update...

See merge request 140crafts/use-jinear!38

+++ jinear-core/src/main/java/co/jinear/core/manager/material/MaterialOperationManager.java
@@ -4,6 +4,7 @@ import co.jinear.core.converter.material.MaterialInitializeFileUploadRequestToVo
 import co.jinear.core.converter.material.MaterialInitializeFolderRequestToVoConverter;
 import co.jinear.core.model.dto.material.MaterialDto;
 import co.jinear.core.model.dto.material.WaitingForUploadMaterialResultDto;
+import co.jinear.core.model.enumtype.material.MaterialAccessType;
 import co.jinear.core.model.request.material.MaterialInitializeFileUploadRequest;
 import co.jinear.core.model.request.material.MaterialInitializeFolderRequest;
 import co.jinear.core.model.request.material.MaterialMoveRequest;
@@ -17,6 +18,7 @@ import co.jinear.core.service.material.MaterialAccessValidationService;
 import co.jinear.core.service.material.MaterialOperationService;
 import co.jinear.core.service.material.MaterialRetrieveService;
 import co.jinear.core.service.passive.PassiveService;
+import co.jinear.core.validator.workspace.WorkspaceMediaLimitQueryService;
 import co.jinear.core.validator.workspace.WorkspaceValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -35,6 +37,7 @@ public class MaterialOperationManager {
     private final MaterialInitializeFileUploadRequestToVoConverter materialInitializeFileUploadRequestToVoConverter;
     private final MaterialRetrieveService materialRetrieveService;
     private final PassiveService passiveService;
+    private final WorkspaceMediaLimitQueryService workspaceMediaLimitQueryService;
 
     public BaseResponse initializeFolder(MaterialInitializeFolderRequest materialInitializeFolderRequest, String workspaceId) {
         String currentAccountId = sessionInfoService.currentAccountId();
@@ -49,6 +52,7 @@ public class MaterialOperationManager {
         String currentAccountId = sessionInfoService.currentAccountId();
         MaterialDto materialDto = materialRetrieveService.retrieve(materialMoveRequest.getMaterialId());
         workspaceValidator.validateHasAccess(currentAccountId, materialDto.getWorkspaceId());
+        materialAccessValidationService.validateMaterialChangeAccess(currentAccountId, materialDto);
         log.info("Move material has started. currentAccountId: {}", currentAccountId);
         materialOperationService.moveTo(materialMoveRequest.getMaterialId(), materialMoveRequest.getParentMaterialId());
         return new BaseResponse();
@@ -58,6 +62,7 @@ public class MaterialOperationManager {
         String currentAccountId = sessionInfoService.currentAccountId();
         MaterialDto materialDto = materialRetrieveService.retrieve(materialRenameRequest.getMaterialId());
         workspaceValidator.validateHasAccess(currentAccountId, materialDto.getWorkspaceId());
+        materialAccessValidationService.validateMaterialChangeAccess(currentAccountId, materialDto);
         log.info("Rename material has started. currentAccountId: {}", currentAccountId);
         materialOperationService.rename(materialRenameRequest.getMaterialId(), materialRenameRequest.getNewName());
         return new BaseResponse();
@@ -66,6 +71,10 @@ public class MaterialOperationManager {
     public MaterialInitializeFileUploadResponse initializeFileUpload(MaterialInitializeFileUploadRequest materialInitializeFileUploadRequest, String workspaceId) {
         String currentAccountId = sessionInfoService.currentAccountId();
         workspaceValidator.validateHasAccess(currentAccountId, workspaceId);
+        workspaceMediaLimitQueryService.validateWorkspaceStorageLimitNotExceeded(workspaceId, materialInitializeFileUploadRequest.getFileSize());
+        if (materialInitializeFileUploadRequest.getParentMaterialId() != null) {
+            materialAccessValidationService.validateMaterialReadAccess(currentAccountId, materialInitializeFileUploadRequest.getParentMaterialId());
+        }
         log.info("Initialize file upload has started. currentAccountId: {}", currentAccountId);
         MaterialFileInitializeVo materialFileInitializeVo = materialInitializeFileUploadRequestToVoConverter.convert(materialInitializeFileUploadRequest, workspaceId, currentAccountId);
         WaitingForUploadMaterialResultDto waitingForUploadMaterialResultDto = materialOperationService.initialize(materialFileInitializeVo);
@@ -85,9 +94,20 @@ public class MaterialOperationManager {
         String currentAccountId = sessionInfoService.currentAccountId();
         MaterialDto materialDto = materialRetrieveService.retrieve(materialId);
         workspaceValidator.validateHasAccess(currentAccountId, materialDto.getWorkspaceId());
+        materialAccessValidationService.validateMaterialChangeAccess(currentAccountId, materialDto);
         log.info("Delete material has started. currentAccountId: {}", currentAccountId);
         String passiveId = passiveService.createUserActionPassive(currentAccountId);
         materialOperationService.delete(materialId, passiveId);
+        return new BaseResponse();
+    }
+
+    public BaseResponse updateAccessType(String materialId, MaterialAccessType materialAccessType) {
+        String currentAccountId = sessionInfoService.currentAccountId();
+        MaterialDto materialDto = materialRetrieveService.retrieve(materialId);
+        workspaceValidator.validateHasAccess(currentAccountId, materialDto.getWorkspaceId());
+        materialAccessValidationService.validateMaterialChangeAccess(currentAccountId, materialDto);
+        log.info("Update access type has started. currentAccountId: {}", currentAccountId);
+        materialOperationService.updateAccessType(materialId, materialAccessType, currentAccountId);
         return new BaseResponse();
     }
